Complexity Refactor: phase 3 (rewriting implementation of Grakn gRPC - Server and Client) (#2840)
# Why is this PR needed?

The way we implemented Grakn GRPC was very convoluted, hard to understand, maintain and extend.

# What does the PR do?

We've re-written the GRPC implementation in `grakn-client` (now `client-java`) and in `grakn-engine`, and completely removed `grakn-grpc`. We've also redefined our GRPC protocol definitions (now `client-protocol`) to have a structure; and one that resembles our Transaction and Concept API and should be easy to maintain and extend. Along the way, we've also improved the usability of Grakn in Java, by refining the Session+Transaction API, as well as Concept API (all the method names). There were many more usability issues and bugs fixed along the way too.

Due to the nature of the work, this PR is massive (sorry guys!). I recommend pulling my branch, and review this in an IDE to get a feeling & understanding of the new architecture. Once you do, there are 3 interesting places to look at:
1. `client-protocol` : new Protobuf API
2. `client-java` : 
    - `ai.grakn.client.concepts.*`
    - `ai.grakn.client.rpc.*`
    - `ai.grakn.client.Grakn`, `.Session` and `.Transaction`
    - (ignore `ai.grakn.batch`, this subpackage will be deleted)
3. `grakn-engine` : `ai.grakn.engine.rpc.*`

If anyone wants to make another drive in any other language, `client-java` can be the reference implementation, which should be simple.

# Does it break backwards compatibility?

Yes, for the GRPC client drivers depending on the old protocol definitions.

# List of future improvements not on this PR

Query `Answer` inference explanation implementation in GRPC, which will follow immediately after this PR.