replaced Transaction.Query.Req.infer from bool to INFER (enum) to workaround GRPC restriction of boolean defaulting false (#2894)
# Why is this PR needed?

GRPC sets the default value of booleans to `FALSE`, when a field is not set by a user. This is an issue for one of our fields in `SessionProto` : `Transaction.Query.Req.infer` which determines whether a query will perform inference or not. Graql (on the server) expects to perform inference (`infer=true`) as the default behaviour if the user has not set the parameter to `false`. However, if GRPC always set `infer=false` when the user doesn't set it, then this misleads the server to think that the user has set it to `false`.

# What does the PR do?

Make the `Transaction.Query.Req.infer` in `SessionProto` to be defined as `enum INFER { TRUE=0; FALSE=1 }`. GRPC sets default values of enums to be the first index, and thus `TRUE` will be the default.