Move delete keyspace functionality from Tx to KeyspaceStore (#2920)
# Why is this PR needed?

Good question.
We currently delete a Keyspace by doing `tx.delete()`, this is a bit counterintuitive for a user who doesnt know all the Grakn details.
Also we currently send few auto-requests to our REST controllers from within the Engine:
 - `EmbeddedGraknSession` although it has `embedded` in the name, it still sends `CommitLog`s via REST using `EngineCommunicator` 🥇 every 1 seconds (via a newly spawned thread)
- `EmbeddedGraknSession` also gets `grakn.properties` as a big string using `EngineCommunicator` 💯 
- When we delete a keyspace using `tx.delete()` we secretly send a self REST DELETE request to `/kb/:keyspace` to delete the keyspace from the list of available keyspaces that we hold in the `grakn-system` keyspace
All the above made sense when we still did not have gRPC and a vision for what goes in the Client and what's on the Server, but know it should all change now.

Also gRPC client has Keyspace deletion on the gRPC transaction.   

On the server side we have a strange entry point to create sessions `Grakn.session()` that uses reflection to invoke an `EmbeddedGraknSession`. This was needed when we still believed in the vendor agnostic thing.

`GraknConfig` is actually a `GraknConfigFactory` that is printing out all the config paths every time someone invokes `GraknConfig.create()`.

`EmbeddedGraknSession` reads `grakn.properties` file (via REST) to get all the necessary information, but for some reason the Engine URI (which is contained in the `grakn.properties`) it is a separate parameter - probably because when set it to `IN_MEMORY` we create and in-memory graph.


# What does the PR do?

- Moves keyspace deletion to `KeyspaceStore` (which needs to be renamed probably) on the server side
- Moves keyspace deletion on `new Grakn().keyspaces().delete()` on the client side

- Deletes `Grakn` static class from server side that was by reflection invoking `EmbeddedGraknSession`. The user should only get a session via gRPC now, and if someone will be embedding Grakn then they will be able to either use the Server to get an `EngineGraknTxFactory` (which caches sessions) or access directly the `EmbeddedGraknSession` class.

- `GraknConfig` now caches the default properties read from `grakn.properties` files, so that it doesnt end up printing hundred times the config paths that are in use
- Removes `delete()` method from `GraknAdmin` which is the admin version of a transaction
- GraqlSession now uses
- Removes `KeyspaceSession` - this was created to accommodate changes from KGMS - but it will need to be rethinked, so removed for now

*EmbeddedGraknSession*

- Exposes new `inMemory()` method on `EmbeddedGraknSession` to create in memory sessions - this replaces the need of passing `IN_MEMORY` value as Engine URI.
- `EmbeddedGraknSession` stores a cache of in-memory sessions in order to make tests pass (they were previously cached in the deleted `Grakn` class), we can probably refactor the tests and delete this in the future
- Removes the method `uri()` from `GraknSession` as it is not used anymore (given that the uri is contained in the configs)

*REST endpoints*
-  Having the keyspace deletion on KeyspaceStore means that finally if a user sends a DELETE REST request to `kb/:keyspace` we actually delete the keyspace, not only remove it from the list of existing ones - yaaay)
- Removes `CommitLogController` and its endpoint from REST endpoints
- Removes endpoint to PUT `/kb/:keyspace` that was used to retrieve `grakn.properties` as a big string
- Deletes `EngineCommunicator`

*EmbeddedGraknTx*

Removed `close(bool, bool)` method and split it into:

`close()` - close tx without committing
`commit()` - commit and closes tx
`commitAndGetLogs()` - commit, closes and compute `CommitLog` which is returned

Now the only ways a user can commit are:
 - via gRPC transaction, which on the backend uses SessionProvider which uses `commitAndGetLogs()` to post `CommitLog` directly to `PostProcessor`
- via REST using the `execute` endpoint, which also uses `commitAndGetLogs()` and post it directly to `PostProcessor`

If someone wants to embed Grakn in his app, he will have acces to `EmbeddedGraknTx` and to the PP, so he will similarly be able to call `commitAndGetLogs()`.

(All the tests use `commit()` method )

Refactored tests and deleted `GraknSessionLocal` class, that was used to simulate a version of Engine capable of reading `grakn.properties` directly from file and not using the self REST request.

# Does it break backwards compatibility?

Yes if someone is relying on `tx.delete()` method.
Also the Grakn client version is not static anymore, it is now an object that needs to be instantiated with `Grakn client = new Grakn(URI)`, on `client` is then possible to call `session(keyspace)` or `keyspaces()` to access keyspace functionalities.
Hopefully nobody was relying on `Grakn.session()` that uses reflection - if so they will need to move to the client or use the `EngineGraknTxFactory` provided by the server


# List of future improvements not on this PR

There are still a lot of things to be changed in regard to Sessions and Transactions, this is only a first step.

NOTE: opening sessions directly through `EmbeddedGraknSession` it will not add the new keyspace to the list of available ones, that's why it should never be used by the user. It is only used in tests right now, one day it should ideally become a private entry point for the Server and not be exposed to the public.

Also one day we should rely directly on a TinkerPop `org.apache.tinkerpop.gremlin.server.GraphManager` to retrieve all existing keyspaces and we should drop a graph using `org.janusgraph.core.JanusGraphFactory` to drop a keyspace, as it will also automatically remove it from the graphManager's list