Merged QueryParser interface and implementation, and migrated the package (#4637)
# Why is this PR needed?

As part of the process of splitting Graql Query classes from to not depend on Transactions (issue #4456), we need to first simplify the Graql Query classes. For this instance, `QueryParser` is an interface with one implementation (and will stay that way). It is a package that is only needed by `graql` and not by `server`, and it should also not depend on `server`.

# What does the PR do?

- Merge `QueryParser` interface and implementation
- Migrate `parser.*` from `grakn.core.graql.internal.parser.*` to `grakn.core.graql.parser.*` and make it not depend on `server`.