Refactor Graql query language library - Part 3 (#4687)
# Why is this PR needed?

This PR is a continuation of PR #4646, and incremental work for issue #4456 where you can find the full description of the goal. In summary, we are refactoring Graql to extract the Graql language library from Grakn's implementation of Graql, and we start with simplifying the Graql codebase first.

# What does the PR do?

This PR is an intermediate step in simplifying the architecture of Graql Query and Pattern classes. We first simplify the unnecessary abstractions to make it easier to see the behaviour of the code.
1. Remove the usage of AutoValue in Property classes, so we can easily change them as we want
2. Remove the usage of AutoValue in *Query classes, so we can easily change them as we want
3. Lots of ad-hoc cleanups to help to understand the code.