New package address and maven coordinate (#2)
## What is the goal of this PR?

Now that client-java sits in an independent repository, it is meant to be generic and works with both grakn-core and grakn-kgms. And like every other repository, it should have its own package address prefix and Maven Group ID: `grakn.client`. Given that the whole repository comprises one maven package, it does not have a _"main"_ package of entry for users to refer to with an Artifact ID. And given that it is an API to grakn-core and grakn-kgms, we've chosen the most appropriate name for the Maven artifact would be `api`. Thus, the full Maven coordinate would be: `grakn.client:api`.

## What are the changes implemented in this PR?

- Replace all package address prefix from `grakn.core.client` to `grakn.client`
- Replaced `maven_coordinates` in `BUILD` file to be: `grakn.client:api:{pom_version}`

Additionally,
- Split CI jobs into two: `build` and `test`
- Renamed GitHub template files to be consistent: uppercase with underscore