Retrieve Explanations Sequentially (#57)
## What is the goal of this PR?
Matching protocol changes in https://github.com/graknlabs/protocol/pull/20, we now iteratively retrieve layers of the `Explanation` tree of a `ConceptMap` answer. `ConceptMap.explanation()` does a round trip to the Grakn server and retrieves a list of `ConceptMap` that explain how the given concept map was obtained. We also restrict `Explanation` to only be obtainable from `ConceptMap`, given that we have not implemented `Explanation` on delete, count, etc. queries.

## What are the changes implemented in this PR?
* Implement https://github.com/graknlabs/protocol/pull/20
* `ConceptMap` contains a `queryPattern` -- this is empty if reasoner is not triggered, and the new method `hasExplanation()` will return `false`
* When `hasExplanation()` is true, it is possible to send the ConceptMap back to the server via `Transaction` to obtain an `Explanation` object
* `Explanation` object only contains a list of `ConceptMap`
* Refactor tests and add an Explanation test