Cast meta types to Type base type so that subs is compatible (#99)
## What is the goal of this PR?

fix #98 

To fix an issue where getting meta types via concept API causes a cast error. The `MetaType` class should be treated as a `Type` so that it is compatible with regular type instances. This should prevents calls like `subs` on the meta type from incorrectly casting their subs to a `MetaType`.

## What are the changes implemented in this PR?

* `MetaType`s are treated like `Type` for the purposes of casting.