Revert to 1.7.x compatible deps and syntax for release (#103)
## What is the goal of this PR?
In preparation of release: client-java was ahead of grakn-core and other clients in that it implemented Delete functionality for 1.8 and used `ValueType` instead of `DataType` (also 1.8.x functionality). This also required it used the next version of the Graql and protocol (though it is possible to use it as gRPC affords flexibility here, we don't want guesswork of which dependency is "ahead" and which is "in sync" in general). We aim to keep 1.N in line between Clients and Core, using the in-sync and compatible Graql and Protocol dependencies.

## What are the changes implemented in this PR?
* Revert dependencies back to 1.0.6 and 1.0.5 for Graql and Protocol respectively
* Relatedly, undo `ValueType` and use `DataType`, and remove one instance of new `delete` syntax