Fix metatypes (#111)
## What is the goal of this PR?

To fix the basic metatype getters on transaction and add tests to ensure they don't break again in future. This was broken because the API was confused between using metatypes and using the actual types. Here, we can switch properly to using the actual types, as they contain all the methods and are type-compatible with `subs()`.

## What are the changes implemented in this PR?

- Metatype getters now get the actual type class, rather than the special metatype class. This is how the implementation worked and the fact that they didn't would just cause a casting exception, so this is a bug fix.