Split Relation.players into 2 methods: one filtered by roles, and one that isn't; and replace Role with RoleType in Thing classes (#150)
## What is the goal of this PR?

To fix Relation.players() always returning an empty collection, and to enforce consistent naming.

## What are the changes implemented in this PR?

In https://github.com/graknlabs/grakn-2.0/pull/13, we updated the Grakn RPC to take in two different kinds of messages for retrieving roleplayers from a relation; one filtered by roles, and one that isn't.

Then we aligned client-java's `Relation` and `RelationImpl` with the new message format.

In doing this we realised that methods named `roles()` returning RoleTypes was weird and unclear, particularly when called from Things. So we decided to refactor these methods.