Upgrade 2.0 build and upgrade nodejs rules (#107)
## What is the goal of this PR?
Update to use cleaned up deployment targets, new format of deployment properties in `deployment.bzl`. We also implement the new style of running release-dependency validation using a bazel test rule instead of a script called in CI. This required adding `workspace_refs` to WORKSPACE.

This PR includes  all changes from https://github.com/graknlabs/client-nodejs/pull/106 .

NOTE: the current state is that client-nodejs is not expected to run or pass tests, as Grakn 2.0 is not finished, and client-nodejs is not updated at all yet with 2.0 changes in protocol etc.