Align package structure, type hierarchy and Concept API with Core (#157)
## What is the goal of this PR?

To make client-java beautiful and functional.

Substantially improving the code quality and usability of client-java by aligning the package structure, type hierarchy and Concept API with Core.

## What are the changes implemented in this PR?

We broke down the 1400-line beast that was GraknClient, separating it into 6 files:

- an interface, named Grakn, containing sub-interfaces named Transaction, Session, Client, Database and DatabaseManager;
- an implementation file for each of the 5 sub-interfaces.

Then, we introduced a method Grakn.client() that instantiates a Grakn.Client. And we separated out some of the inner classes of Transaction, including GraknTransceiver and RpcIterator, while refactoring out QueryOptions completely in favour of the Options class that Grakn Core uses.

We aligned Concept API and its type hierarchy with Core, deleting, renaming and adding methods as appropriate; deleting `MetaType` in favour of `ThingType` (for 'thing'), `AttributeType` (for 'attribute'), and so on. We also deleted almost all generics from Concept API, which made the internal code more awkward, but the user-facing code cleaner. We also refactored out `Label` in favour of `String`, and `ConceptIID` in favour of another `String`.

Then, we refactored out the `rpc` package by moving Transceiver to `connection`, RequestBuilder to `connection.ConnectionMessageWriter` and ResponseReader to `answer.AnswerMessageReader`. We also refactored methods out of `ResponseReader` and into the `Answer` subtypes, to align the structure of `answer` with `concept`.

After that we introduced the `Concepts` interface into `concept`, made an implementation called `GraknConcepts` that takes in a `GraknTransceiver`, and refactored the entire `concept` package to only ever reference the `Concepts`, not a `Transaction`. By doing so, `concept` package no longer has a dependency on the root package.

We refactored out `GraknConceptException` and unified all Exceptions into `GraknClientException`, constructed using an `ErrorMessage` which extends the `ErrorMessage` in the **common** repository - similar to the way Core does it. This means that each exception thrown on the Client has a unique error code.