Implement missing race condition fix from master (#165)
## What is the goal of this PR?

We patched a race condition that caused transactions to hang during closing a while ago but the patch was never released and only exists in master. During debugging of a new issue, we rediscovered the same problem, but are now more sure about the nature of the bug and a more correct fix. However, we are also certain that the original fix has proven itself to be stable, so in the interest of keeping master the same as the latest release, we are going to release the original fix as 1.8.3.

## What are the changes implemented in this PR?

- Add `synchronized` to transceiver methods to ensure that no calls result in a race condition.