Basic CI Testing (#117)
## What is the goal of this PR?

While presently client-nodejs is functional and tested, the tests presently have to be run manually whenever we wish to check if they've been passed, which is easy to forget to do. To this end, this PR seeks to integrate code quality assurance into our automated CI through basic E2E tests in the test phase of automation.

## What are the changes implemented in this PR?

We introduced extremely basic end to end tests, testing the simple cases of each individual function of the client.
We integrated these tests (again, in a very simple sense) with our automated CI system.
In addition to the above simple E2E tests, we introduced a linter to enforce proper code style and likewise integrated it into CI.
Finally, we made the necessary modifications to the code to actually pass said linter, as it turns out we were not yet adhering to best practises.