Fix errors being swallowed + update protocol (#119)
## What is the goal of this PR?

Session pulses were incorrectly swallowing errors. We are making them throw errors as normal, and furthermore, they actually read the server response now and close the session locally if the session is marked as closed remotely.

Also, gRPC errors were being similarly swallowed. We are making them shut down the transaction and inform all response listeners that an error occurred.

We also had a nasty solution to circular dependency issues that was misaligned with client-java: the `ConceptProtoReader`. We are dissolving it in favour of putting the static `of` methods into namespaces whose names match the class name.

Also, the `npm` artifact of protocol was renamed to `grakn-protocol`.

## What are the changes implemented in this PR?

- Session pulses now close the session locally if the session is marked as closed remotely, and errors are no longer swallowed
- gRPC errors now shut down the transaction and inform all response listeners that an error occurred
- Dissolved `ConceptProtoReader`
- Import `grakn-protocol`, not `graknlabs-protocol`