Protocol Synchronisation (#128)
## What is the goal of this PR?

We synchronised client-nodejs's conceptmanager and concept type structure with the newest protocol for grakn core.

## What are the changes implemented in this PR?

getType on the concept manager is now private, and correctly fetches a thingtype (and is named to reflect this), per the new protocol. Additionally, as the base "thing" is now returned by getSupertypes again, the type signatures of getSupertypes have been updated to reflect this.