API cleanup: close on commit, isX/asX methods (#216)
## What is the goal of this PR?

As we close transactions on commit at the server end, it makes sense to also close transactions when they are committed at client side. This change does that. Additionally, for user experience reasons, "isX" functions have been added to the Concept interface to determine what variety of Concept an object represents.

## What are the changes implemented in this PR?

We added a try/finally to `commit()` that closes the transaction when it completes or fails.
We added is[Concept] functions to all Concepts indicating if the current Concept is an instance of that Concept variety.