Close duplex streams when closing transactions (#144)
## What is the goal of this PR?

Tests were experiencing intermittent failures due to improper cleanup of transactions. We cleared up the cleanup steps so they're easier to read and less temperamental, and added a step to transaction.close() that actually closes the RPC stream so the tests do not hang sporadically.

## What are the changes implemented in this PR?

We cleaned up connectionsteps in the behavioural tests.
We added _stream.end() to the process of closing transactions.