Supports Grakn Cluster failover functionality (#232)
## What is the goal of this PR?

We have added support for Grakn Cluster failover, allowing the client to switch between talking with one server to another. With this functionality, the users can now benefit from Cluster's high-availability and horizontal scalability.

Instantiating a client for Grakn Cluster is similar to instantiating a client for Grakn Core. When it connects to the server, it will retrieve complete information about the whole cluster and other servers that belong there.

```
GraknClient.Cluster client = new GraknClient.Cluster("172.0.0.1:1729");
```

When connected to Grakn Cluster, the client and session instances are bound to the whole cluster rather than a particular server. That means that if one server in the cluster crashes, the client and session instances are still fully operational and will simply failover to other servers that are available.

Transaction instances however, are bound to the server they connect to and cannot failover to other servers midway. Therefore if a server crashes, the transactions will simply fail.

Apart from the failover ability, the client behaviour when connecting to Grakn Cluster is otherwise identical to how it behaves when connecting to Grakn Core.

## What are the changes implemented in this PR?

Grakn Cluster specific implementations of the following component:

- `GraknClient.Cluster`: a client implementation that is designed to work with a cluster instance. It contains an array of `GraknClient.Core`, one for each server in the cluster
- `RPCSession.Cluster`: a session implementation that is designed to work with a cluster instance. It contains an array of `RPCSession.Core`, one for each server in the cluster.
- `RPCDatabaseManager.Cluster`: a database manager implementation for managing databases clusters