Add new package for Cluster and delete Grakn interface (#248)
## What is the goal of this PR?

We had Cluster-specific code living in the same classes as Grakn Core code. We made a conscious decision to view Cluster features as an extension of Core features, as the Cluster product is an extension of the Core product. To reflect this shift in mentality we created a new package named `rpc.cluster`. All Cluster functionality (besides that which cannot be easily moved, such as error messages) now lives there.

We also got rid of the Grakn interface, as it was a little odd having a method named `GraknClient.core` that returned `Grakn.Client`. The Client interface is now `GraknClient` and the Session, Transaction and DatabaseManager now live under `GraknClient` (e.g: `GraknClient.Transaction`)

## What are the changes implemented in this PR?

- Move Cluster-specific code to rpc.cluster package
- Delete Grakn and use GraknClient as the root interface for the main classes (Session, Transaction etc)