Fix client-java throwing exception when multiple queries run in parallel inside one transaction (#278)
## What is the goal of this PR?

We fixed an issue where queries run in parallel inside one transaction were not executed successfully and would throw an exception instead.

## What are the changes implemented in this PR?

According to `StreamObserver` documention:

```
Separate StreamObservers do not need to be synchronized together; incoming and outgoing directions are independent. Since individual StreamObservers are not thread-safe, if multiple threads will be writing to a StreamObserver concurrently, the application must synchronize calls.
```

Since the same `StreamObserver` is used for all the queries inside one transactions, it was causing issues. We fixed it by wrapping `StreamObserver` in our own helper class `SynchronizedStreamObserver` that enforces all the calls to `onNext` and `onCompleted` to be `synchronized`.

close #275