Recover faster from temporary network outages and server failures (#285)
## What is the goal of this PR?

In the event of a network interruption, the client would previously start "automatically" rejecting all requests for a certain period of time. We now force the client to try to reconnect if the user performs any operation, meaning that if the server is restored after a temporary outage, it will now be available immediately to process fresh requests.

## What are the changes implemented in this PR?

After any server outage, always try to reconnect when performing any operation

## Additional information

This PR is based off https://github.com/graknlabs/client-java/pull/284 and should be merged only once that PR is.