Grakn Cluster integration (#165)
## What is the goal of this PR?

We now support connecting to Grakn Cluster using
```js
client = await GraknClient.cluster(addresses)
```
where `addresses` is an array of Cluster server addresses. The API usage of a Cluster client is the same as for a Core client, however, the creation of a Cluster client returns a `Promise<GraknClient>` (unlike creating a Core client, which returns a `GraknClient`). It also supports one additional field in `GraknOptions` named `readAnyReplica`, which when set to `true` on a `Session`, allows all of its **read** transactions to read from secondary database replicas.

## What are the changes implemented in this PR?

Create Cluster-specific versions of GraknClient, Session, Database, and DatabaseManager for use when connecting to Cluster.