Architecture Refactor (#302)
## What is the goal of this PR?

Over time, Client Java has been extended considerably to implement various features and optimisations. It has lead the architecture to a complex codebase which is not so easy to follow and replicate in other languages. We have now refactored Client Java considerably to come up with an architecture that is easy to understand, easy to extend, and easy to replicate in other languages that wish to implement a Grakn Client.

1) Made the Transaction Bidirectional Stream architecture a first class citizen of the architecture, making it obvious and easier to understand and follow. The Bidirectional Stream now gets its own package at `//stream`.
2) Made the two main responsibilities of the bidirectional stream - single response and streamed multipart response - to be explicitly enforced and expressed.
3) Simplified the response parsing logic across the codebase from passing around lambdas to eventually be applied at the lowest level (at "ResponseIterator"), by reversing the logic such that every lower-level class returns the top-level response type and every top level class will unpack the response before returning upwards.
4) Introducing the the "API" package in which all interfaces are aggregated into. This allows us to remove cycles across the classes that was complicated to follow in other language implementations of Grakn Client.
5) Split the monolithic repository (there was only one "package" that compiled all source code) into multiple packages, enforcing and expressing a clear hierarchy. 

The diagram below describes the new architecture of client-java, which should be able to be replicated across every other language implementation of Grakn Clients.

![package-structure](https://user-images.githubusercontent.com/567679/111183012-61191b00-85a7-11eb-935d-95edc8cbb177.png)