Serialised the closing of request dispatchers (#312)
## What is the goal of this PR?

We forgot to protect the request dispatcher (`RequestTransmitter.Dispatcher`) from sending messages on the GRPC `StreamObserver` after it's being closed. Given that query operations in a transaction is no fully asynchronous, closing a transaction early (without committing) may result in those query operations being sent after the transaction is closed. It should not cause any bugs to the application but it would cause the background thread that executes the request dispatcher to be throwing exceptions. We have now fixed this problem by introducing an atomic boolean to determine the state of the request dispatcher, and we've made the `close()` function `synchronized`.