Improve TypeDB Cluster support: Introduce TLS encryption and improve database management operations (#328)
## What is the goal of this PR?

We have added two major improvements, namely:

### Brand-new TLS encryption support

We've added TLS network encryption support for secure communication with TypeDB Cluster allowing for safe transmission of sensitive data.

### Improved database management

We have removed the workaround for TypeDB Cluster's database management operations. Prior to this PR, the client executes the operation on each server in the cluster.

In this PR, the client only goes to one server and let that server coordinate the operation to the other servers. This is made possible by leveraging the new database management coordination feature that has been implemented in TypeDB Cluster.

## What are the changes implemented in this PR?

### TLS encryption
- Added TLS network encryption support with configurable root CA.

### Database management 
- `create`, `get`, and `contains` calls go to any replica, but may optionally go to the primary replica.
- `delete` call must go to the primary replica