API refactors for ease of use (and yarn!) (#184)
## What is the goal of this PR?

We've added an `index.ts` file so that symbols are importable from "typedb-client" as opposed to, say, "typedb-client/api/connection/TypeDBTransaction".

Also, we've added casting methods to Concept classes (e.g: `asThingType`). These methods perform safe type casting.

We've aligned the class/namespace structure fully with client-java - previously, for example, `AttributeType.Boolean.Remote` was flattened to `AttributeType.RemoteBoolean`. This is no longer the case.

We've found `yarn` is faster and more deterministic than `npm` for installing dependencies, so we've updated the build system to use `yarn`. This does not affect users.

(Also, the generated lockfile is **much** smaller!) 

## What are the changes implemented in this PR?

- Shorten import paths
- Add Concept type cast methods
- Flatten class structures
- Use `yarn` instead of `npm` to install dependencies