Don't delete response collectors in a transaction (#369)
## What is the goal of this PR?

We no longer delete response collectors in a transaction after receiving a response to a "single" request, or receiving a "DONE" message in a stream. This fixes a possible error when loading 50+ answers in one query and then performing a second query.

## What are the changes implemented in this PR?

See https://github.com/vaticle/typedb-client-python/pull/250, which this PR is a copy of.