Fix memory leak in long-running sessions (#389)
## What is the goal of this PR?

A long running session used to keep the reference to each closed transaction, continually growing in size. These references are inaccessible from outside the session object. The problem was compounded in the write transactions, when the caller did not collect the responses received after insert queries, which were held onto by the closed transaction objects.

We now dispose of the closed transaction objects and free their memory.

## What are the changes implemented in this PR?

Transaction now keeps a backreference to the session that spawned it and notifies it when it is closed. The session releases the reference to the transaction and allows the GC to free the memory.