TypeDB Cluster support (#27)
## What is the goal of this PR?

We add limited TypeDB Cluster support to the client-rust. A cluster
client can open a connect with or without TLS, manage databases, open
transactions and send queries. User management is not yet supported.

## What are the changes implemented in this PR?

Pre-existing TypeDB core functionality is moved into the new
`connection::server` module, with the exception of `TypeDBClient`, which
is relocated to `connection::core`. We then introduce a new
`connection::cluster` module that defines the cluster connection API.

Conceptually, `cluster` API deals with the cluster as a whole, `core`
expects TypeDB Core, and `server` can interface with either a core
server or a single cluster node. This convention extends into the
`common::rpc` module as well.

The failover mechanism from client-java is implemented on
`cluster::Database` instead. This allows us to avoid shared ownership as
much as possible, as all the data that is needed for failovers —info on
database replicas— is available at that point.

Closes #9.