Check BDD value equality using native types (#417)
## What is the goal of this PR?

In BDD steps implementation we checked the equality of all values through the conversion to `String`s. It might be incorrect when comparing `Double`s especially if one of these values is a result of arithmetics. 
Now we compare `Double`s by their absolute error. 
For consistency we compare all values using their native types equality.

## What are the changes implemented in this PR?

- We modified `ValueUniquenessCheck` class in order to compare values using their native types equality.
- We modified `KeyUniquenessCheck` class: we find the desired key attribute label and compare each type immediately instead of converting all values to `String`.