Introduce protocol versioning (#421)
## What is the goal of this PR?

We use a new protocol API to perform a "connection open". This API does server-side protocol version compatibility checks, and replaces our previous need to get all databases to check that the connection is available.

The user will receive an error about a protocol version mismatch if they are using a client-server combination that are not using exactly compatible protocols. 
1. The server will raise an error if the client tries to connect with a mismatching protocol version
2. The client will raise an error if it tries to connect to the server and the server does not have that API

Both errors imply a client-server mismatch and the error will suggest this as a fix.

This change depends on https://github.com/vaticle/typedb-protocol/pull/185, which means that this client will implement protocol version 1.

## What are the changes implemented in this PR?

* Refactor `validateConnectionOrThrow` into `openConnection`, was then inlined into the constructors of `CoreClient` and `ClusterServerClient`
* Call the new `openConnection` with the protocol Version set when constructing a client
* Throw a useful message if the server returns an `unimplemented` error